NumericWithLabel with Style
- goal details contains more data
- most data fades on collapse

+++ app/src/main/java/au/com/beba/runninggoal/feature/goals/WorkoutViewHolder.kt
@@ -19,7 +19,7 @@ class WorkoutViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
 
     fun bindView(workout: Workout) {
         val dt = GoalDate(workout.dateTime)
-        val d = Distance.fromMetres(workout.distanceInMetres.toLong())
+        val d = Distance.fromMetres(workout.distanceInMetres)
 
         name.text = workout.name
         description.text = workout.description
